refactor(state): converted config module into state module
