+++ lua/ashen/setup.lua
@@ -2,7 +2,7 @@ local M = {}
 
 M.colors = function()
   local colors = require("ashen.colors")
-  local opts = require("ashen.config").opts
+  local opts = require("ashen.state").opts
   local override = opts.colors or {}
   for k, v in pairs(override) do
     colors[k] = v
